pyproject.toml
[project]
name = "thsl"
description = "A typed configuration language"
authors = [
    {name = "Anthony Post", email = "anthony.post@icloud.com"},
]
dependencies = [
    "python-dateutil==2.8.2",
    "tempora==4.1.1",
    "semantic-version==2.8.5"
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
version = { source = "file", path = "thsl/__version__.py"}

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.0.240",
    "mypy>=0.991",
    "pytest>=7.2.1",
    "ufmt>=2.0.1",
]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.mypy]
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
pretty = true
exclude = ["__pypackages__", "tests"]

[[tool.mypy.overrides]]
module = [
    "semantic_version",
    "tempora"
]
ignore_missing_imports = true

[tool.ruff]
select = ["ALL"]
ignore = [
    "D", "C", "I", "PD", "ANN101", "ANN102", "A003", "EM101", "EM102", "TRY003",
    "FBT001", "FBT002", "TCH003", "PGH003", "ERA001", "E999",
]

#fixable = []
#unfixable = []

exclude = [
    "tests",
]

[tool.ruff.per-file-ignores]
